+++ index.php
@@ -5,7 +5,7 @@ if (!$connection['link']) {
     $page_content = include_template('error.php',
         ['error' => $connection['error']]);
 } else {
-    $categories = getCategories($connection['link'], $error);
+    $categories = get_сategories($connection['link'], $error);
     if ($categories) {
         $page_content = include_template('main.php',
             ['categories' => $categories]);
@@ -13,7 +13,7 @@ if (!$connection['link']) {
         $page_content = include_template('error.php', ['error' => $error]);
     }
 
-    $lots = getActiveLots($connection['link'], $error);
+    $lots = get_active_lots($connection['link'], $error);
     if ($lots) {
         $page_content = include_template('main.php', [
             'categories' => $categories,
