+++ init.php
@@ -2,10 +2,15 @@
 date_default_timezone_set("Europe/Moscow");
 $config = require 'config/config.php';
 require_once('helpers.php');
-require_once('function.php');
+require_once('functions/getters/db.php');
+require_once('functions/getters/lot.php');
+require_once('functions/database/db.php');
+require_once('functions/validators/lot.php');
+require_once('functions/function.php');
+
 require_once('data.php');
 
-$connection = dbConnect($config['db']);
+$connection = db_connect($config['db']);
 
 $error='';
 $categories = [];
