+++ pages/login.html
@@ -57,7 +57,7 @@
         </li>
       </ul>
     </nav>
-    <form class="form container" action="https://echo.htmlacademy.ru" method="post"> <!-- form--invalid -->
+    <form class="form container" action="https://echo.htmlacademy.ru" method="post"> <!-- getters--invalid -->
       <h2>Вход</h2>
       <div class="form__item"> <!-- form__item--invalid -->
         <label for="email">E-mail <sup>*</sup></label>
