+++ pages/sign-up.html
@@ -57,7 +57,7 @@
         </li>
       </ul>
     </nav>
-    <form class="form container form--invalid" action="https://echo.htmlacademy.ru" method="post" autocomplete="off"> <!-- form
+    <form class="form container form--invalid" action="https://echo.htmlacademy.ru" method="post" autocomplete="off"> <!-- getters
     --invalid -->
       <h2>Регистрация нового аккаунта</h2>
       <div class="form__item"> <!-- form__item--invalid -->
