+++ templates/add-lot.php
@@ -10,7 +10,7 @@
             <label for="lot-name">Наименование <sup>*</sup></label>
             <input id="lot-name" type="text" name="lot-name"
                    placeholder="Введите наименование лота" <?= 'value="'
-            .getPostVal('lot-name').'"'; ?>>
+            .get_post_val('lot-name').'"'; ?>>
             <span class="form__error"><?= $errors['lot-name'] ?></span>
         </div>
         <?php $classname = isset($errors['category']) ? "form__item--invalid"
@@ -22,7 +22,7 @@
                 <?php foreach ($categories as $category): ?>
                     <option
                         <?= 'value="'.$category['id'].'"'; ?>
-                        <?php if ($category['id'] == getPostVal('category')): ?>
+                        <?php if ($category['id'] == get_post_val('category')): ?>
                             selected
                         <?php endif; ?>
                     ><?= $category['name']; ?>
@@ -37,7 +37,7 @@
     <div class="form__item form__item--wide  <?= $classname; ?>">
         <label for="message">Описание <sup>*</sup></label>
         <textarea id="message" name="message"
-                  placeholder="Напишите описание лота"><?= getPostVal('message'); ?></textarea>
+                  placeholder="Напишите описание лота"><?= get_post_val('message'); ?></textarea>
         <span class="form__error"><?= $errors['message'] ?></span>
     </div>
     <?php $classname = isset($errors['file']) ? "form__item--invalid"
@@ -45,8 +45,7 @@
     <div class="form__item form__item--file  <?= $classname; ?>">
         <label>Изображение <sup>*</sup></label>
         <div class="form__input-file   <?= $classname; ?>">
-            <input class="visually-hidden" type="file" id="lot-img"
-                   name="lot_img">
+            <input class="visually-hidden" type="file" id="lot-img" name="lot_img">
             <label for="lot-img">
                 Добавить
             </label>
@@ -59,7 +58,7 @@
             <label for="lot-rate">Начальная цена <sup>*</sup></label>
             <input id="lot-rate" type="text" name="lot-rate"
                    placeholder="0" <?= 'value="'
-            .getPostVal('lot-rate').'"'; ?>>
+            .get_post_val('lot-rate').'"'; ?>>
             <span class="form__error"><?= $errors['lot-rate'] ?></span>
         </div>
         <?php $classname = isset($errors['lot-step']) ? "form__item--invalid"
@@ -68,7 +67,7 @@
             <label for="lot-step">Шаг ставки <sup>*</sup></label>
             <input id="lot-step" type="text" name="lot-step"
                    placeholder="0" <?= 'value="'
-            .getPostVal('lot-step').'"'; ?>>
+            .get_post_val('lot-step').'"'; ?>>
             <span class="form__error"><?= $errors['lot-step'] ?></span>
         </div>
         <?php $classname = isset($errors['lot-date']) ? "form__item--invalid"
@@ -78,7 +77,7 @@
             <input class="form__input-date" id="lot-date" type="text"
                    name="lot-date"
                    placeholder="Введите дату в формате ГГГГ-ММ-ДД" <?= 'value="'
-            .getPostVal('lot-date').'"'; ?>>
+            .get_post_val('lot-date').'"'; ?>>
             <span class="form__error"><?= $errors['lot-date'] ?></span>
         </div>
     </div>
